requirements.txt
opencv
streamlit
Pillow
